+++ lua/todo-sidebar/scanner.lua
